refactor: Separate test context (#56)
