+++ core/core-api/src/functionalTest/resources/application-functional-test.yml
