+++ core/core-api/src/integrationTest/groovy/com/hbc/pms/core/api/IntegrationTestSpec.groovy
@@ -1,18 +1,14 @@
-package com.hbc.pms.core.api.test.setup
+package com.hbc.pms.core.api
 
-import com.hbc.pms.core.api.CoreApiApplication
-import com.hbc.pms.core.api.TestDataFixture
 import com.hbc.pms.support.spock.test.AbstractFunctionalSpec
-import org.spockframework.spring.EnableSharedInjection
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
-import spock.lang.Shared
 import spock.util.concurrent.PollingConditions
 
 @SpringBootTest(classes = [CoreApiApplication], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-class FunctionalTestSpec extends AbstractFunctionalSpec {
+@ActiveProfiles("integration-test")
+class IntegrationTestSpec extends AbstractFunctionalSpec {
   PollingConditions conditions = new PollingConditions(timeout: 20)
 
   @Autowired
@@ -21,7 +17,8 @@ class FunctionalTestSpec extends AbstractFunctionalSpec {
   def setup() {
     dataFixture.populate()
   }
-  def cleanup(){
+
+  def cleanup() {
     dataFixture.cleanup()
   }
