+++ core/core-api/src/integrationTest/resources/container-license-acceptance.txt
@@ -1,0 +1,1 @@
+mcr.microsoft.com/mssql/server:2017-CU12
