+++ gradle.properties
@@ -17,3 +17,4 @@ plc4xVersion=0.11.0
 systemProp.sonar.projectKey=f-pms_API
 systemProp.sonar.organization=f-pms
 systemProp.sonar.host.url=https://sonarcloud.io
+systemProp.sonar.coverage.jacoco.xmlReportPaths=build/reports/jacoco/testCodeCoverageReport/testCodeCoverageReport.xml
